requirements.txt
adlfs==2024.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
azure-core==1.32.0
azure-datalake-store==0.0.53
azure-identity==1.19.0
azure-storage-blob==12.24.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
duckdb==1.1.3
frozenlist==1.5.0
fsspec==2024.12.0
idna==3.10
isodate==0.7.2
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
numpy==2.2.1
pandas==2.2.3
portalocker==2.10.1
propcache==0.2.1
pyarrow==18.1.0
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websockets==14.1
yarl==1.18.3
